refactor: replace configVars by hostSpec
In order to add work machine to this config some modularity is required.
hostSpec aims to replace the clunky way of providing user variable to
the whole nix-config.

The home-configuration was also dropped because it wasn't used at all
and add too much constraints to the whole config wihtout any benefit.

Inspired by EmergentMind's nix-config.