refactor: use the module system instead of optional imported part
Adding new bits of configuration is becoming painful as new hosts are
added. The configuration needed to become more composable.

To make the configuration more composable, the import of optional module
were replaced by the usage of nix's options.

Although this adds a lot of indirection is also came with the benefits
of adding a lot of flexibility which will be needed to add more hosts.

The architecture used here is highly inspired on raf's nix-config
https://github.com/NotAShelf/nyx.git